+++ ext/hash/php_hash_sha.h
@@ -40,7 +40,7 @@ typedef struct {
 } PHP_SHA1_CTX;
 
 PHP_HASH_API void PHP_SHA1Init(PHP_SHA1_CTX *);
-PHP_HASH_API void PHP_SHA1Update(PHP_SHA1_CTX *, const unsigned char *, unsigned int);
+PHP_HASH_API void PHP_SHA1Update(PHP_SHA1_CTX *, const unsigned char *, size_t);
 PHP_HASH_API void PHP_SHA1Final(unsigned char[20], PHP_SHA1_CTX *);
 
 PHP_FUNCTION(sha1);
@@ -56,7 +56,7 @@ typedef struct {
 } PHP_SHA224_CTX;
 
 PHP_HASH_API void PHP_SHA224Init(PHP_SHA224_CTX *);
-PHP_HASH_API void PHP_SHA224Update(PHP_SHA224_CTX *, const unsigned char *, unsigned int);
+PHP_HASH_API void PHP_SHA224Update(PHP_SHA224_CTX *, const unsigned char *, size_t);
 PHP_HASH_API void PHP_SHA224Final(unsigned char[28], PHP_SHA224_CTX *);
 
 /* SHA256 context. */
@@ -67,7 +67,7 @@ typedef struct {
 } PHP_SHA256_CTX;
 
 PHP_HASH_API void PHP_SHA256Init(PHP_SHA256_CTX *);
-PHP_HASH_API void PHP_SHA256Update(PHP_SHA256_CTX *, const unsigned char *, unsigned int);
+PHP_HASH_API void PHP_SHA256Update(PHP_SHA256_CTX *, const unsigned char *, size_t);
 PHP_HASH_API void PHP_SHA256Final(unsigned char[32], PHP_SHA256_CTX *);
 
 /* SHA384 context */
@@ -78,7 +78,7 @@ typedef struct {
 } PHP_SHA384_CTX;
 
 PHP_HASH_API void PHP_SHA384Init(PHP_SHA384_CTX *);
-PHP_HASH_API void PHP_SHA384Update(PHP_SHA384_CTX *, const unsigned char *, unsigned int);
+PHP_HASH_API void PHP_SHA384Update(PHP_SHA384_CTX *, const unsigned char *, size_t);
 PHP_HASH_API void PHP_SHA384Final(unsigned char[48], PHP_SHA384_CTX *);
 
 /* SHA512 context */
@@ -89,7 +89,7 @@ typedef struct {
 } PHP_SHA512_CTX;
 
 PHP_HASH_API void PHP_SHA512Init(PHP_SHA512_CTX *);
-PHP_HASH_API void PHP_SHA512Update(PHP_SHA512_CTX *, const unsigned char *, unsigned int);
+PHP_HASH_API void PHP_SHA512Update(PHP_SHA512_CTX *, const unsigned char *, size_t);
 PHP_HASH_API void PHP_SHA512Final(unsigned char[64], PHP_SHA512_CTX *);
 
 PHP_HASH_API void PHP_SHA512_256Init(PHP_SHA512_CTX *);
